(RAZOR-145) Add dmidecode (and virt-what) to microkernel
Without this patch the microkernel 004 release includes `dmidecode`, but
the build automation tools do not explicitly express a dependency on
dmidecode.  This is a problem because it is not clear there is a
requirement to always include `dmidecode` in the microkernel.  If this
requirement is not met for some reason in the future key facts may be
omitted.

This patch addresses the problem by explicitly adding dmidecode to the
kickstart configuration for the microkernel in an effort to ensure the
executable is always present for use with facter.

In addition, virt-what is added since this tool helps provide more
accurate virtualization fact support and so this dependency is
explicitly listed in the kickstart configuration.